Revisited the visitor pattern :^)
To be able to return different types from a 'visit' many things had to be changed.
A visit method now doesn't return anything, instead the visitor itself stores the result of the last visit.
To make everything a bit easier, I also introduced four macros to visit and fetch the result in one call.